Transactions : Corrected the "Amount" column in the "Transactions" tab
Made changes so that the "Amount" column will always show the transaction amount with the transaction fee for outgoing transactions and just the amount for incoming transactions.

Reason for changes : There was a bug that caused the amounts displayed under the "Transactions" tab to sometimes include the fee and sometimes not include the fee for outgoing transactions. This was seen when the same wallet file was used with SimpleWallet in the command line to send a transaction and then the same wallet file was opened using the GUI wallet. These outgoing transactions did not include the fee in the "Amounts" column. Also, if the wallet was restored from the private keys in the GUI wallet, the amounts displayed were correct.