Transaction Details Box : Corrected the amounts sent and received
Reason for changes : When a user double clicks on a transaction under the "Transactions" tab, a transaction details dialog box will pop up. In this box, the previously displayed amount sent and received was off by 0.01 CASH2 which is the current transaction fee. The changes made here correct the amounts shown in the pop up box.